Split up PriceCalculator block and component
